+++ src/Zeus/Control/Exception/ControlException.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace GetOlympus\Zeus\Control\Exception;
+
+use GetOlympus\Zeus\Base\Exception\BaseException;
+
+/**
+ * Control exception.
+ *
+ * @package    OlympusZeusCore
+ * @subpackage Control\Exception
+ * @author     Achraf Chouk <achrafchouk@gmail.com>
+ * @since      2.0.2
+ *
+ */
+
+class ControlException extends BaseException
+{
+    //
+}